requirements.txt
streamlit>=1.25.0
Pillow>=10.0.0
ultralytics>=8.0.100